requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
marshmallow>=3.20.1
pytest>=8.0.0
Sphinx>=7.0.0
pytest-cov>=4.0.0
